C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik4.cpp -o C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik4.exe  -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wmissing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -D_DEBUG -D_EJUDGE_CLIENT_SIDE -D_EJC

+++ Ozhaev Zaynudin Sharik4.cpp
@@ -116,8 +116,8 @@ void MoveBall()
       		e = Dy/d; // cos
       		Vn1 = vx2*s + vy2*e;
       		Vn2 = vx1*s + vy1*e;
-      		int Vt1 = -vx2*e+vy2*s;
-      		int Vt2 = -vx1*e+vy1*s;
+      		float Vt1 = -vx2*e+vy2*s;
+      		float Vt2 = -vx1*e+vy1*s;
 
 
       		float o = Vn2;
